Show line number tooltip on the right if side panel is collapsed
